feat: add basic request handling
