+++ frontend/src/components/OsaForm.tsx
@@ -35,7 +35,6 @@ export default function OsaForm() {
 
   const onSubmit: SubmitHandler<FormType> = async (data) => {
     console.log({ data });
-    console.log({ errors });
     try {
       await fetch("/api/submit", {
         method: "POST",
